+++ src/gdal_common.hpp
@@ -89,6 +89,10 @@ std::shared_ptr<RETURN> NumberArrayToSharedPtr(Local<Array> array, size_t count 
 
 #define ATTR(t, name, get, set) Nan::SetAccessor(t->InstanceTemplate(), Nan::New(name).ToLocalChecked(), get, set);
 
+#define ATTR_ASYNCABLE(t, name, get, set)                                                                              \
+  Nan::SetAccessor(t->InstanceTemplate(), Nan::New(name).ToLocalChecked(), get, set);                                  \
+  Nan::SetAccessor(t->InstanceTemplate(), Nan::New(name "Async").ToLocalChecked(), get##Async, READ_ONLY_SETTER);
+
 #define ATTR_DONT_ENUM(t, name, get, set)                                                                              \
   Nan::SetAccessor(t->InstanceTemplate(), Nan::New(name).ToLocalChecked(), get, set, Local<Value>(), DEFAULT, DontEnum);