+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
  - GDAL 3.3.1
+ - Support async getters
+   - Add `gdal.Dataset.rasterSizeAsync`
+   - Add `gdal.Dataset.srsAsync`
+   - Add `gdal.Dataset.geoTransformAsync`
  - Add `gdal.DatasetBands.getEnvelope()`
  - Add `gdal.vsimem.set` and `gdal.vsimem.release` allowing to load and unload files between a `Buffer` and GDAL's in-memory `vsimem` file system
  - Add `gdal.Geometry.makeValid{Async}()`